# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PESAMATRIX SIGNAL API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface RegisterInput {
  email: string;
  password: string;
  name: string;
}

export interface ChangePasswordInput {
  currentPassword: string;
  newPassword: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  ADMIN: 'ADMIN',
  USER: 'USER',
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  mustChangePassword: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type SignalDirection = typeof SignalDirection[keyof typeof SignalDirection];


export const SignalDirection = {
  BUY: 'BUY',
  SELL: 'SELL',
} as const;

export type SignalStatus = typeof SignalStatus[keyof typeof SignalStatus];


export const SignalStatus = {
  ACTIVE: 'ACTIVE',
  CLOSED: 'CLOSED',
  PENDING: 'PENDING',
} as const;

export interface Signal {
  id: number;
  pair: string;
  direction: SignalDirection;
  entryPrice: number;
  stopLoss: number;
  takeProfit: number;
  status: SignalStatus;
  /** @nullable */
  pips: number | null;
  createdAt: string;
}

export interface SignalSummary {
  totalSignals: number;
  winRate: number;
  activePlan: string;
  totalProfitPips: number;
  weeklyChange: number;
  winRateChange: number;
}

export type SubscriptionStatus = typeof SubscriptionStatus[keyof typeof SubscriptionStatus];


export const SubscriptionStatus = {
  ACTIVE: 'ACTIVE',
  EXPIRED: 'EXPIRED',
  PENDING: 'PENDING',
  CANCELLED: 'CANCELLED',
} as const;

export interface Subscription {
  id: number;
  userId: number;
  status: SubscriptionStatus;
  daysSelected: number;
  totalAmount: number;
  feePerDay: number;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  createdAt: string;
}

export interface SubscriptionInput {
  daysSelected: number;
  phoneNumber?: string;
}

export type PaymentStatus = typeof PaymentStatus[keyof typeof PaymentStatus];


export const PaymentStatus = {
  PENDING: 'PENDING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
  CANCELLED: 'CANCELLED',
} as const;

export interface Payment {
  id: number;
  userId: number;
  /** @nullable */
  subscriptionId?: number | null;
  amount: number;
  status: PaymentStatus;
  method: string;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  checkoutRequestId?: string | null;
  /** @nullable */
  mpesaReceiptNumber?: string | null;
  /** @nullable */
  reference?: string | null;
  /** @nullable */
  failureReason?: string | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface StkPushInput {
  /** Safaricom phone number in format 2547XXXXXXXX */
  phoneNumber: string;
  /** Number of subscription days */
  daysSelected: number;
}

export interface StkPushResponse {
  checkoutRequestId: string;
  merchantRequestId: string;
  paymentId: number;
  message: string;
}

export type PaymentStatusResponseStatus = typeof PaymentStatusResponseStatus[keyof typeof PaymentStatusResponseStatus];


export const PaymentStatusResponseStatus = {
  PENDING: 'PENDING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
  CANCELLED: 'CANCELLED',
} as const;

export interface PaymentStatusResponse {
  status: PaymentStatusResponseStatus;
  paymentId: number;
  /** @nullable */
  mpesaReceiptNumber?: string | null;
  /** @nullable */
  failureReason?: string | null;
  subscription?: Subscription;
}

export interface SystemConfig {
  feePerDay: number;
  minDays: number;
}

export interface SystemConfigInput {
  feePerDay?: number;
  minDays?: number;
}

export interface PerformancePoint {
  day: string;
  value: number;
}

export interface DashboardSummary {
  totalSignals: number;
  winRate: number;
  activePlan: string;
  totalProfitPips: number;
  weeklySignalChange: number;
  winRateChange: number;
  performanceData: PerformancePoint[];
}

export interface CallbackAck {
  ResultCode: number;
  ResultDesc: string;
}

export interface AdminSummary {
  totalUsers: number;
  activeSubscriptions: number;
  totalRevenue: number;
  pendingPayments: number;
  recentUsers: User[];
}

export type Mt5ConnectionStatus = typeof Mt5ConnectionStatus[keyof typeof Mt5ConnectionStatus];


export const Mt5ConnectionStatus = {
  CONNECTED: 'CONNECTED',
  SYNCING: 'SYNCING',
  DISCONNECTED: 'DISCONNECTED',
  ERROR: 'ERROR',
} as const;

export interface SlaveAccount {
  id: number;
  userId: number;
  mt5Login: string;
  brokerServer: string;
  status: Mt5ConnectionStatus;
  /** @nullable */
  statusMessage: string | null;
  /** @nullable */
  metaApiAccountId: string | null;
  /** @nullable */
  lastSyncAt: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ConnectMt5Input {
  mt5Login: string;
  mt5Password: string;
  brokerServer: string;
}

export interface UpdateMt5Input {
  mt5Password?: string;
  brokerServer?: string;
}