# lib/api-zod/src/generated/types/connectMt5Input.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PESAMATRIX SIGNAL API
 * OpenAPI spec version: 0.1.0
 */

export interface ConnectMt5Input {
  mt5Login: string;
  mt5Password: string;
  brokerServer: string;
}