# lib/api-zod/src/generated/types/mt5ConnectionStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PESAMATRIX SIGNAL API
 * OpenAPI spec version: 0.1.0
 */

export type Mt5ConnectionStatus = typeof Mt5ConnectionStatus[keyof typeof Mt5ConnectionStatus];


export const Mt5ConnectionStatus = {
  CONNECTED: 'CONNECTED',
  SYNCING: 'SYNCING',
  DISCONNECTED: 'DISCONNECTED',
  ERROR: 'ERROR',
} as const;