# lib/api-zod/src/generated/types/slaveAccount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PESAMATRIX SIGNAL API
 * OpenAPI spec version: 0.1.0
 */
import type { Mt5ConnectionStatus } from './mt5ConnectionStatus';

export interface SlaveAccount {
  id: number;
  userId: number;
  mt5Login: string;
  brokerServer: string;
  status: Mt5ConnectionStatus;
  /** @nullable */
  statusMessage: string | null;
  /** @nullable */
  metaApiAccountId: string | null;
  /** @nullable */
  lastSyncAt: string | null;
  createdAt: string;
  updatedAt: string;
}