# lib/api-zod/src/generated/types/updateMt5Input.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PESAMATRIX SIGNAL API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateMt5Input {
  mt5Password?: string;
  brokerServer?: string;
}